refactor * Update, Write 컴포넌트 통합 * parameter 여부에 따라 적용
