--- frontend/src/Components/Write/Update.module.css
@@ -1,153 +1,0 @@
-.board_top h2 {
-    width: 1200px;
-    font-size: 40px;
-    font-weight: bold;
-}
-
-.board_top h2,
-.form {
-    display: flex;
-    flex-direction: column;
-    width: 800px;
-    margin: 1rem auto;
-}
-
-.mbti_checkbox {
-    display: flex;
-    flex-direction: column;
-    margin-top: 30px;
-    margin-bottom: 10px;
-    align-items: center;
-}
-/* MBTI 선택  */
-.mbti {
-    display: flex;
-    flex-wrap: wrap;
-    height: 80px;
-    align-items: center;
-    justify-content: center;
-}
-
-.mbti div input[type='checkbox'] {
-    display: none;
-}
-.mbti div input[type='checkbox'] + label {
-    width: 100px;
-    font-weight: 400;
-    border-radius: 5px;
-    padding-top: 2px;
-}
-.mbti div input[type='checkbox'] + label:hover {
-    cursor: pointer;
-}
-
-.mbtibox {
-    width: 90px;
-    line-height: 30px;
-    color: #fff;
-}
-
-.mbtibox:hover {
-    transition: 0.3s;
-    font-weight: 700;
-    transform: scale(1.1);
-}
-
-.mbti div {
-    display: flex;
-    justify-content: center;
-    flex: 0 0 12.5%;
-    text-align: center;
-    user-select: none;
-    transition: transform 0.3s ease-in-out;
-}
-
-/* 제목 */
-.title,
-.content {
-    margin-top: 1rem;
-    width: 100%;
-}
-
-.title div {
-    height: 2.5rem;
-    display: flex;
-    justify-content: space-between;
-}
-
-.title div input {
-    width: 60%;
-    height: 100%;
-    border-radius: 20px;
-    border: 1px solid gray;
-}
-
-.title div select {
-    margin-right: 10px;
-    width: 40%;
-    font-size: 1rem;
-    padding-left: 10px;
-    border-radius: 20px;
-    border: 1px solid gray;
-}
-
-.title input {
-    font-size: 1rem;
-    height: 2rem;
-    padding-left: 20px;
-}
-
-.title select:hover {
-    cursor: pointer;
-}
-
-.content div {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-}
-
-/* 내용 */
-.content textarea {
-    font-size: 1rem;
-    height: 20rem;
-    white-space: pre-wrap;
-    border-radius: 30px;
-    padding: 20px;
-}
-
-/* Error mesage */
-.Error {
-    margin: 15px auto;
-    color: red;
-    height: 10px;
-    font-weight: bold;
-}
-
-/* 버튼 */
-.button {
-    height: 50px;
-    width: 500px;
-    margin: 5px auto;
-    background-color: rgba(201, 201, 201, 79%);
-    border: 0 solid;
-    border-radius: 10px;
-    font-size: 1.2rem;
-}
-
-.button:hover {
-    background-color: rgba(170, 170, 170, 79%);
-    transition: 0.2s;
-    cursor: pointer;
-}
-
-.toggleSwitchContainer {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    margin-top: 5px;
-}
-
-.toggleSwitchContainer > h3 {
-    margin-right: 5px;
-}
